+++ HISTORY.md
@@ -21,6 +21,7 @@
 
 ### 2021-07-07
 * morphic, gui: tweaked perish() animation
+* objects: fixed "hide / show primitives" for unified palette
 
 ### 2021-07-06
 * byob, objects: new feature: hide custom helper blocks in palette